+++ demo2blindNFT/scripts/discard/mint.js
