+++ demo2blindNFT/scripts/discard/reveal.js
@@ -1,14 +1,15 @@
 
 // 开启盲盒脚本
+
 require("@nomiclabs/hardhat-ethers");
 const hre = require("hardhat");
-const config = require("../../hardhat.config");
+const config = require("../hardhat.config");
 
 // NFT合约部署成功后的地址
 const CONTRACT_ADDRESS = config.mintConfig.contractAddress;
 
 // 合约abi接口
-const contractInterface = require("../../artifacts/contracts/TripNFT.sol/TripNFT.json").abi;
+const contractInterface = require("../artifacts/contracts/TripNFT.sol/TripNFT.json").abi;
 
 // 直接构建 infura provider
 const provider = new hre.ethers.providers.InfuraProvider(config.mintConfig.network, config.mintConfig.infuraKey);
@@ -25,7 +26,7 @@ const contractWithSigner = contract.connect(wallet);
 // 入口函数
 async function main() {
     // 开启盲盒
-    await contractWithSigner.flipReveal();
+     await contractWithSigner.flipReveal();
 }
 
 // We recommend this pattern to be able to use async/await everywhere
