+++ demo2blindNFT/scripts/freemint.js
