+++ demo2blindNFT/scripts/publicmint.js
