+++ demo2blindNFT/webapp/package.json
@@ -13,6 +13,8 @@
     "count-time-down": "^1.0.5",
     "env-cmd": "^10.1.0",
     "ethers": "^5.7.1",
+    "keccak256": "^1.0.6",
+    "merkletreejs": "^0.2.32",
     "next": "12.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
